pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "groq>=1.2.0",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.15",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
